Created a new folder for the new 4th iteration of the PCB
I need to adjust the SPI Header to allow for the reset signal to be sent, along with this I will just route an extra pin from the mega to the touch SPI Interface, just in case someone wants to take advantage of that in the future. This also makes it so we can use a standard 2x6 connector and not waste any pins.